renamed HDF5Management
